Allow some env to be set outside config
I wanted to temporarily set a different version and local host name
without editing the config file, so I thought it would be easier to
allow these environment variables to be set outside the configuration
script. The change allows them to use whatever was set in the
environment, and default to the config file if not set.

Also, add checksum for Hadoop 3.0.2