Added instruction in the docs/centos7.md file (#213)
Added instructions to make the authorized_keys file read/write for user so ssh to localhost will work. It is set 664 by default with the instructions we give.  This will cause ssh localhost to  prompt for passphase.  If you don't check the ssh setup setup after running the commands in centos.md and go right to the Quickstart commands ssh won't work. The command uno setup accumulo will let you think everything went well but it did not.